# fix: try re-execute statement if backend responds with "prepared statement is no longer valid"