--- a/pgjdbc/src/main/java/org/postgresql/jdbc/AutoSave.java
+++ b/pgjdbc/src/main/java/org/postgresql/jdbc/AutoSave.java
@@ -1,0 +1,21 @@
+package org.postgresql.jdbc;
+
+public enum AutoSave {
+  NEVER,
+  ALWAYS,
+  CONSERVATIVE;
+
+  private final String value;
+
+  AutoSave() {
+    value = this.name().toLowerCase();
+  }
+
+  public String value() {
+    return value;
+  }
+
+  public static AutoSave of(String value) {
+    return valueOf(value.toUpperCase());
+  }
+}