… pgjdbc/src/main/java/org/postgresql/core/v3/SimpleQuery.java
… pgjdbc/src/main/java/org/postgresql/ds/common/BaseDataSource.java
… pgjdbc/src/main/java/org/postgresql/jdbc/BatchResultHandler.java
… pgjdbc/src/main/java/org/postgresql/jdbc/CallableBatchResultHandler.java
… pgjdbc/src/main/java/org/postgresql/jdbc/PgConnection.java
… pgjdbc/src/main/java/org/postgresql/jdbc/PgResultSet.java
… pgjdbc/src/main/java/org/postgresql/jdbc/PgStatement.java
… pgjdbc/src/main/java/org/postgresql/util/PSQLState.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc2/AutoRollbackTestSuite.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc2/PreparedStatementTest.java
… ubenchmark/src/main/java/org/postgresql/benchmark/statement/ParseStatement.java