Bug report: http://stackoverflow.com/questions/34180932/error-cached-plan-must-not-change-result-type-when-mixing-ddl-with-select-via
When using "deallocate all" or "alter table", cached prepared statements might become invalid, thus we need to reexecute query in those cases

It might require some logging, however I'm not sure what is the best way to log that.

Summary (technical details):
1) By default pgjdbc watches for `DEALLOCATE ALL`, and `DISCARD ALL` statements, and invalidates prepared statement cache if those are detected.
That does not mean "deallocate all" should be used often. It just means deallocate all might be suitable for maintenance-like procedures.
Note: `deallocate all` invalidates only a single connection.
2) If either `prepared statement XXX does not exist` or `cached statement must not change result type` observed, pgjdbc would invalidate prepared statement cache (see `deallocateEpoch++;`). In general those errors should not happen, however if observed it means something had changed behind the scenes.
3) By default pgjdbc would retry `prepared statement XXX does not exist` and `cached statement must not change result type` errors in `autocommit=true` (non-transactional) mode.
4) New connection property: `autosave`
 * `conservative` -- rollback to savepoint only in case of "prepared statement does not exist" and
       "cached plan must not change result type". Then the driver would re-execute the statement ant it would pass through
 * `never` (default) -- never set automatic safepoint. Note: in this mode statements might still fail with "cached plan must not change result type"
       in autoCommit=FALSE mode
 * `always` -- always rollback to "before statement execution" state in case of failure. This mode prevents "current transaction aborted" errors.
       It is similar to psql's ON_ERROR_ROLLBACK.

For `autocommit=true` the savepoint is not used.
In `conservative` mode, savepoint is used only for queries that return something (see `org.postgresql.core.v3.QueryExecutorImpl#sendAutomaticSavepoint`)