@vlsi 2016-08-13
>When such situation detected DBA execute DEALLOCATE ALL but without success

@vladimirfx , Can you elaborate on that?
AFAIK, `deallocate all` should be done through pgjdbc connection. How does DBA do that?