@vlsi 2016-08-13
Unfortunately, using SAVEPOINTS and DML causes extra xid allocation, so I do not see good solution for "cached plan cannot change result type" in transactional mode (see https://github.com/pgjdbc/pgjdbc/pull/477#issuecomment-239644770)

deallocate/discard can be solved without savepoints (the driver would just notice deallocate responses and invalidate the cache).