@davecramer 2016-08-15
I don't think this is a good reason to abandon AutoSave feature. the psqlODBC driver has it. Also I didn't realize that any select that does NOT use a transaction uses up an XID as well. As I think most people don't use transactions they are consuming XID's at a considerable rate without complaining.