@vlsi 2016-08-15
Some field test is required to check how autosave=conservative impacts xid allocation rate in real applications.