@davecramer 2016-08-15
Well unfortunately the only way we are going to get field tests is by enabling the feature and getting reports back. I think we put the feature in. Do not make it the default. If people want the feature they can enable it.