@vlsi 2016-08-15
> I think we put the feature in. Do not make it the default.

Ok good. I was having the same in mind.


Just brainstorming for other possible solutions: we can try LISTEN/NOTIFY to cooperate pgjdbc with maintenance procedures.
For instance:
1) Before DB maintenance (e.g. alter table, etc), DBA issues `NOTIFY 'pgjdbc', 'please, use unnamed statements only'`.
2) As far as I understand, every backend would receive that NOTIFY message, and pgjdbc would eventually consume it. There's a glitch that "newly established connections would not know the maintenance is in progress".
3) As pgjdbc consumes that message, it will go into "unnamed statements" mode.
4) DB upgrade is done. Note: since no long-lived statements are in use, the probability of "cached plan must not change result types" kind of errors should be minimal
5) As the maintenance completes, DBA issues `NOTIFY 'pgjdbc', 'ok, you might use named statements again'`.