@iqbalmd 2016-08-18
@davecramer 
How to get field tests is by enabling the feature and getting reports back??