@vlsi 2016-08-19
@iqbalmd , you can try 9.4.1210-SNAPSHOT version: https://github.com/pgjdbc/pgjdbc#development-snapshots.
You can grab the snapshot jar from maven central:  https://oss.sonatype.org/content/repositories/snapshots/org/postgresql/postgresql/9.4.1210-SNAPSHOT/

Automatic savepointing is activated via `autosave=conservative` or `autosave=always` connection setting.