@santtuh 2016-08-19
I tested the snapshot version and it fixes my (simulated) test case. Eagerly waiting for this fix to be released.