@vladimirfx 2016-09-14
@vlsi 
> AFAIK, deallocate all should be done through pgjdbc connection. How does DBA do that?
Actually him not classic DBA, operational engineer. He call JDBC command in Apache Karaf console, it may use same pool with application. It is just the try to fix first error ("cached statement cannot change result type") adhoc.
We try new release and give you feedback. Thanks a lot for the fix!