@laglace 2016-10-11
Vsli,

I'm currently in the middle of a Oracle -> Postgres Migration Process, and I have to support both database systems in our application. Finding out that the Postgres Database Rolls Back Transactions upon an error made me stumble upon the "Autosave" feature you implemented, so I promptly installed the newest version of the Driver.

We're using JBoss and XA-Datasources, so I simply added a new xa-datasource-property, but alas, I could not get it working! (Only "AutoSave" was considered a valid property, and it promptly sent a "method .setAutoSave() does not exist exception". I analyzed the code and, in fact, I could only find the ".setAutosave()" Method)

In the end, I had to cast our Connection Object to a PGConnection Object inside the application, which works as intended, but is a solution we'd rather avoid. 

Can you confirm this behaviour?