@vlsi 2016-10-11
> Only "AutoSave" was considered a valid property

I do not follow that. Can you try "autosave"?

Have you tried to add the property via url parameter? (see https://github.com/pgjdbc/pgjdbc#jdbc-url-syntax )
E.g. `jdbc:postgresql://host:port/database?autosave=always`