@mj1618 2018-03-04
Yea so here are our numbers, I realise how arbitrary they are without the code we used to generate them.

Without autosave conservative:
Avg: 600ms
Min: 50ms
Max: 1800ms
Rps: 130

With autosave conservative:
Avg: 650ms
Min: 35ms
Max: 17100ms
Rps: 120

In this one the worst case was 9.5x worse.

The graphs of these over time looked similar, but the autosave one was bumpy. Ie at some points perf was impacted a lot, but then otherwise it was quite similar.

These weren’t out production soak/stress tests, nor were they run for an extended period, but they were enough to turn us off this solution to the cached query error.